+++ CHANGES.rst
@@ -3,6 +3,9 @@
 
 - Updated maker utility and factory for dark ref model to include group keywords from exposure. [#66]
 
+- Updated maker utilities for level 1, level 2, and ramp models to reflect changes in reference pixels. [#65]
+
+
 0.9.0 (2022-02-04)
 ==================